migrating to .hpp file
